Controller: add python task path to syspath
If a task python is ran and it has some dependency on a module located
in the same directory as the task then the run will fail because the
PYTHONPATH does not contain the task directory.

The patch appends the task directory to the system path variable before
loading python task code and removes it from it afterwards.

An example:

$ ls ~/my_recipes
task.py
ciphers.py
$ cat ~/my_recipes/task.py
from lnst.Controller.Task import ctl

import ciphers
...

$ lnst-ctl run ~/my_recipes/task.py
...
      File "/root/lnst/lnst/Controller/NetTestController.py", line 664, in _run_python_task
        module = imp.load_source(name, task_path.resolve())
      File "/root/testplans/my_recipes/task.py", line 21, in <module>
        import ciphers
    ImportError: No module named ciphers

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>