NetTestSlave: redo mapping of devices that change netns
This patch adds a new slave method "unmap_if" that removes the mapping
of the specified interface id.

I also changed how interface mapping and network namespaces work -
previously the mapping was copied from the main process when the
namespace was created, now the mapping gets cleared and each interface
is mapped separately when it's moved to its target netns and unmapped
when it's returning.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>