Machine: deconfiguration fixes
Interface configuration starts with mapping the interface id to a kernel
interface index. This mapping was previously only globally cleared by
the slave when new configuration started. This patch makes the
controller unmap interfaces one by one in the correct order.

One other minor change - restoring the use_nm option now happens before
calling the "bye" method and not is not called in the "finally" block
since the connection might not be functional anymore.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>