InterfaceManager: keep information about peered interfaces
Device object now know when they're paired with another interface (veth)
this is needed to properly deconfigure them when deconfiguration is not
coordinated by the controller.

This patch also adds the get_id_by_if_index method to the
InterfaceManager which is a helper function that implements the reverse
mapping of if_index to if_id.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>