ConnectionHandler: reimplementation
This commit reimplements the ConnectionHandler class while kkeping the
same interface. The main point was to store file handles in a list
instead of a dictionary which can have name conflicts. The old
dictionary is still used but only to map a connection name to the
connection. This allows us to introduce more mappings in derived classes
and keep the connection checking code from the main class.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>