ServerHandler: inherit from ConnectionHandler
This patch makes the class ServerHandler a class derived from the
ConnectionHandler. In addition to the standard connection_mapping it
adds a new _netns_con_mapping that maps network namespaces to Pipe
objects.

This solves the problems of the previous implementation where we would
have two ConnectionHandlers controlled by the ServerHandler which caused
either 100% CPU consumption or slowdowns because of select timeouts.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>