lnst-ctl: add multi match option
This patch adds a new option [-u|--multi-match] to lnst-ctl that makes
each recipe run with every possible pool match. This can be usefull for
writing recipes that specify a small set of requirements - these can
then be run in many different machine setups by just adding this one
argument.

The patch also modifies the NetTestResultSerializer so that these
matches are logged properly, and the result summary contains information
about the match.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>