SlavePool: small object construction fixes
This patch fixes some small problems in the SlavePool methods
responsible for Machine object construction.

When the machine isn't virtual it doesn't need the libvirt_domain
parameter. This can could cause exceptions when the match wasn't virtual
but the machine had the parameter defined and lnst-ctl wasn't running
with root privileges.

Proper initialization of the network parameter for interfaces. This is
purely so that it doesn't cause confusion in the future. Unused
interfaces don't have any network.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>