TaskAPI: improved access to hosts and interfaces
The ControllerAPI class now supports returning all configured hosts in a
dictionary with this format:
{host_id: HostAPI instance, ...}

A new class InterfaceAPI was added, instances of this class represent
configured interfaces of a host and at the moment can be used for read
access of various interface attributes. In the future this could be
extended to changing these values as well.

HostAPI now has 2 new methods: get_interface and get_interfaces that
return an InterfaceAPI object or a dictionary of {interface_id:
InterfaceAPI object} respectively. The get_* methods that accessed
interface attributes now use the associated InterfaceAPI object while
keeping backwards compatibility.

ModuleAPI gained two new methods: get_options and set_options that make
it possible to reuse the same object multiple times with different
options, e.g. running the IcmpPing module with different intervals.

The ValueAPI class was renamed to VolatileValue and reimplemented to
support all the derived classes out of the box. The objects are
constructed by supplying a function that resolves the desired value with
some arguments. The value can be accessed by calling the get_value
method. This will always call the specified function to get an
up-to-date value. To keep backwards compatibility the __str__ function
is still defined and transforms the value returned by the get_value
method to a string.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>