NetConfigDevice: set bond down before configuration
When NM is running on a machine and lnst is using normal configuration
the bond is created, NM sets it up which makes the configuration of the
bonding mode fail. This patch makes sure that before configuring any
options the device is down. It's a temporary workaround until the issue
is fixed in NM.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>