result_xslt: updated with new information from result xml
This patch updates the XSLT transformation schme to reflect the new
information available in result xmls. This means that the HTML files now
also contain the information about the match number and what pool match
was used.

In addition to that the formatting and functionality was updated. To
make it easier to look at each command (command description and result
data) now has a border, the start of a new Task is highlighted.

There is now a button to show/hide result data of commands. Commands
that don't have any result data don't have this button and instead show
a note saying that no result data was returned. In case of background
commands that return their result data in a corresponding wait/intr/kill
command I added a highlight button that highlights this command, to make
it easier to find.

This patch breaks backwards compatibility and result html files
generated before this patch will now be displayed correctly. To fix this
you can download the old .css and .js files and fix the html file to
reference those instead.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>