NetTestResultSerializer: adjust xml_to_html transformation
This patch works around some issues with the old version of lxml library
available in RHEL6.

The first change is parsing the xslt document from a URL needs to
happen first otherwise the parser is unable to access URLs - know bug in
the old lxml library that is fixed in newer versions.

The second is that the tostring() method didn't support a doctype
parameter in older versions so instead we add the DOCTYPE manually.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>