rename run-git-test, add rsync option, change -g option
This commit renames the bash script that runs lnst regression tests to
run-test.sh. The reason behind this is that the tests don't really have
anything in common with git.

This commit also adds the argument -s which makes the script use rsync
to download lnst from a specified URL instead of using git clone on that
url.

When using the script to run all tests, before execution they will be
numerically sorted so that we keep the ordering in which they were
added.

Finally I changed the -g option (for git url) to -u (for url) since it
makes more sense like this.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>