regression-tests: update test 3
This test contained a mistake - lnst-ctl execution always failed because
no match could be found in a pool - the recipe required a machine with a
parameter and we always created pools with machines that didn't have any
parameters. However the test only checked that lnst-ctl failed and
reported success. This is now fixed.

Additionaly the functionality we tested is the same as with the removed
test 01 and is now supported as a feature. So now the test checks that
lnst-ctl succeeds.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>