XmlProcessing: handle invalid recipe exception
When lnst-ctl reads an invalid recipe xml file it would crash with an
exception, ignoring any other recipes that should have been tested.

This commit fixes that issue by handling the exception and instead of
ending the controller, setting the recipe result to FAIL and continuing
to the next recipe if posible.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>