NetTestCommand: signal handling for bg cmds in netns
Background commands in a network namespace inherited signal handlers
from the lnst-slave process that manages that namespace, this resulted
in wrong handling of the interrupt command for the bg command.

This commit fixes these issues by changing the signal handlers after the
bg cmd fork to raise a KeyboardInterrupt. It also changes the default
SIGINT handler (which was deprecated for a long time) to also raise a
KeyboardInterrupt. Because of this the wait_on_interrupt method was
modified and the read/write was removed from the TestGeneric class.

Finally the commit also adds a new regression test for this bug. The
test consists of running two recipes that differ in the type of the
background command. The first one is a 'run command' that runs an
infinite ping and the seconf one is a 'run module' that runs a modified
version of the IcmpPing module for an infinite ping. The modification is
simple - run the ping twice, this tests if we interrupt the bg cmd
process as well as the exec-ed command.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>