lnst-ctl: add option --dump-config
This commit adds a new option for the LNST controller - --dump-config.
When this option is specified, lnst-ctl will run until all configuration
files are loaded and then dump the resulting configuration on stdout.
This is useful for debugging purposes, or when including lnst in
scripts.

Since this commit the message "Loading config file ..." is printed on
stderr so that it doesn't get mixed with the dumped configuration.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>