regression-tests: don't use user conf by default
This commit changes the default behaviour of running regression tests to
use the configuration present in the regression-tests/env/ directory -
this means that you need to properly create a pool in the env/pool/
directory before running these regression tests.

The option -c was added for situation when you want to use your own
local LNST configuration.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>