ConnectionHandler: import IPRSocket directly to fix isinstance check
Recent change in pyroute2 introduced partial proxy to postpone the import,
on the downside this broke isinstance() check within LNST that did
comparison of unresolved IPRSocket object. More details in
https://github.com/svinota/pyroute2/issues/109

To do the check we can import the class directly from the lower pyroute2
module as suggested by Pete Saveliev.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>