NetTestCommand: fix signal handling for bg commands
This commit fixes signal handling (mainly SIGINT) for background
commands. The default action is to ignore the signals SIGHUP, SIGINT and
SIGTERM. This is needed for "run command" commands to work, and if the
developer of a test module wants to change the behaviour he cna do so in
the run method of his module.

This commit also changes the method wait_on_interrupt provided for test
modules. Previously it was bound to usage coupled with the
set_handle_intr signal handler. Now the method is standalone changes
your SIGINT signal handler only for the duration of the method. This
makes the method set_handle_intr obsolete so it was removed.

Finally the regression test #25 was modified to reflect these changes.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>