Machine: ensure interface deconfiguration if configure was called
This commit moves the "self._configured = True" assignment to the top of
the Interface configure() methods, right after the check is it's already
set. This ensures that if the configuration raised an exception the
deconfigure method will still be called. This ensures that there are no
internal data structures leftover after the failed configuration.

The regression test #26 was added to test for this bug.

This fixes #106.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>