NetTestCommand: use custom signal handler function instead of SIG_IGN
The exec syscall resets all signal handlers except for SIG_IGN. This
means that exec_cmd() doesn't work with <intr> commands. The original
regression test didn't catch this because ping sets it's own SIGINT
handler. To fix this we will use our own handler function that will just
ignore the signal, it has the same behaviour as SIG_IGN, but will get
reset on exec syscall, which is what we want.

The patch also adds a new recipe to the regression test.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>