SlavePool: fix interface parameter matching
Interface compatibility check in matching algorithm compares the value
of parameter only if it's also present in slavemachine definition inside
pool. If it's not in the pool, the check is finally evaluated as matched.
So, if the first interface that we compare to does not have the requested
param and second one has it, it is the first one that is evaluated as
matched and not the second one as a user would expect.

Since we __require__ that the machine has all the requirements we define
under <params> tag in recipe the condition had to be changed.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>