SlavePool: add hostname and hwaddr to match info
This adds hostname and interface hardware addres information to the
dictionary describing the used mapping. This required a minor change to
the interface mapping which is why there are changes in
NetTestController and NetTestResultSerializer as well.

This information is going to be useful when looking at results later
for example in PerfRepo.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>