add lnst/Controller/PerfRepo.py
PerfRepo.py is a module that wraps the REST API provided by a PerfRepo
instance. Classes PerfRepoObject, PerfRepoTest, PerfRepoMetric,
PerfRepoTestExecution, PerfRepoValue and PerfRepoReport define data
objects that we can store or retrieve from the PerfRepo database. They
implement methods for setting and getting their attributes and a helper
functions for xml serialization, they can also be initialized from an
xml-serialized object.

The class PerfRepoRESTAPI actually wraps the REST API provided by
PerfRepo.

This is still an experimental implementation so there are some TODOs,
methods that are just placeholders for now and bugs are expected. What
is implemented is what is needed for basic integration with LNST.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>