Task, Interface: add {get, set}_mtu methods
The Device class (Slave) now tracks mtu changes for the interface and
the changes are sent to the Controller through an update message.

The Interface class (Controller) got two new methods: {get, set}_mtu.
get_mtu just reads the value currently stored in the object (this gets
updated automatically).
set_mtu sends a config command, this means that deconfiguration will
take care of resetting the mtu value to the original, however it also
means that for 'get' we use Netlink and for 'set' we use sysfs. This
will need to be addressed in the future.

Finally the InterfaceAPI class gained the {get, set}_mtu methods that
connect to the Interface class.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>