recipes: swap order of offload settings
With some drivers like ixgbe or e1000e setting tso offload on will return
with non-zero exit code if the offload is already turned on. This makes
the whole recipe fail. The simplest solution to this is reverse the
order in which the offload is set. Setting offload off does not fail if
already set on a device so this change should work.

Signed-off-by: Jan Tluka <jtluka@redhat.com>