NetTestSlave: add get_if_data rpc method
This patch adds a new method 'get_if_data' to the SlaveMethods class.
It returns information about a mapped interface, it will be used after
interface configuration to fill out missing information on the
Controller side.

This also changes the format of 'if_update' messages sent from the Slave
to the Controller, that now sends the return value of the get_if_data
method.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>