Machine: add method get_configuration
This method generates a description of the slave machine that the object
represents, this includes information about configured Interfaces of
this machine.

The commit also updates the ControllerAPI method get_configuration, that
now uses the Machine get_configuration method to create a complete
configuration description of the entire test environment.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>