NetTestSlave: fix traceback for down of nonexistent device
Instead of crashing with a traceback we just log a warning message
saying that the device wasn't found, for the set_unmapped_device_down
method.

For the regular set_device_down method we log an error message and
return False since this probably influences the results of the test.
This still won't stop the test from running but it should at least be
more visible.

Fixes #122

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>