Task: improve generate_hash method
The hash now includes the testUid of the Test that the TestExecution
object is associated with.

The method also accepts a new parameter - ignore which is a list of
parameter patterns, if a parameter matches one of these patterns then
it's skipped and not included into the final hash. This can be used to
skip the redhat-release and kernel-release parameters.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>