add PerfRepoUtils
This module will contain helper functions for PerfRepo integration.
The first two functions are:
parse_id_mapping(filename)
get_id(mapping, key)

These will be used to work with mapping files that are required for full
support of PerfRepo. The plan is to have this functionality in a single
class, but for now these two functions are sufficient.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>