NetTestController: iterate only over StaticInterfaces to get driver info
This is a fix for corner case bug introduced by recent commit
92b8ae0dc8ac25a10835d9898172980b3af89ec1.

The problem is that if the match results with at least one
UnusedInterface we end up iterating over such interface that
unfortunately does not have self._id set since it defaults to
None.

The solution is to query driver information for StaticInterface
interface classes only.

Fixes #131

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Acked-by: Ondrej Lichtner <olichtne@redhat.com>