change shebangs to python2
According to PEP 0394:
in preparation for an eventual change in the default version of Python,
Python 2 only scripts should either be updated to be source compatible
with Python 3 or else to use python2 in the shebang line.

It's not a complete solution but it solves problems of running LNST on
Arch Linux until more work is done on issue #105.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>