InterfaceManager: check ethtool output
In case ethtool output is not matched by the regex the match object is
None and an exception will be raised on the match.group(1) call. This
patch fixes that by instead returning None (based on a patch from
Artem Savkov).

The issue can appear if the slave machine doesn't have ethtool installed
which doesn't normally happen (because of dependency resolution) but can
appear when lnst is installed/running from the git version.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>