Netperf: more refactorization
This patch brings more refactorization to the Netperf test module.
Instead of parsing options multiple times in different functions all the
options are parsed in the __init__ method of the class. Except for the
"netperf_server" option there's no need to check if the slaves role is
client or server so all of the options are parsed in the same place.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>