Netperf: add support for confidence (-I option)
This patch adds support for the confidence feature of Netperf. You can
start using this feature by defining the 'confidence' option of the test
module. The format is identical to what is accepted by the netperf
client so:
lvl[,intvl]
where:
  lvl     is either 95 or 99 and specifies the confidence level
  intlvl  is the width of the confidence interval as a percentage

You can further configure the min and max number of runs by specifying
the -i option using the 'netperf_opts' test module option.

The resulting measured confidence is used to compute the deviation of
the result and creating the interval that will be used for comparison
against a threshold.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>