recipes: add support for PerfRepo to phase1 tests and refactoring
This patch adds initial support for PerfRepo integration to the phase1
regression test recipes. All recipes with PerfRepo integration have an
updated README file that describes how to enable this feature. Please
remember that you also need to configure the LNST Controller to be able
to connect to a PerfRepo instance.

In addition to that I refactored the usage of test modules by reusing
the same object with updated parameters for multiple runs of the same
test.

The patch also contains a minor fix connected to the test duration
alias - changing the test duration wasn't reflected in the timeout value
of the test therefore resulting in timeouts in case you defined a high
enough value in the alias.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>