NetTestCommand: add {get, set}_result_sent method
This commit adds the {get, set}_result_sent methods, and a new attribute
(_result_sent) to the NetTestCommand class. This attribute is used to
indicate that the result of a command was already sent to the
Controller. The set method takes an optional parameter 'value' that can
be used to set the attribute to False if needed (not currently used).

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>