NetTestSlave: check if command finished before killing
This commit makes shure that when the rpc method kill_command is called,
it first checks if the command didn't finish first. If it did, it will
only be removed from the command context. This situation can occur when
the Controller detects a command timeout and sends a kill_command
request, but on the Slave the command is already finished and the result
was sent in the mean time.

This commit should fix this issue and the command will be seen as
finished on time. If the command was killed we now also add a killed
flag to the result message to indicate this action.

Related to #120

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>