regression_tests: fix generate_hash ignore list
When generating the hash of a result we were supposed to ignore the
parameters "kernel_release" and "redhat_release", instead we were
ignoring "kernel-release" and "redhat-release". This patch fixes the
issue.

In case of virtual tests I also added two more regex-es that ignore the
hostname and mac addresses of the guest machines as these will almost
certainly be different between runs since we always generate new guests.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>