PerfRepo: reimplement Report properties handling
Until now we only supported getting baselines from the Report
properties, and the implementation was far from ideal.

This patch removes the parsing and moves to working directly with the
dictionary structure, which is a bit better than what we had, and also
adds implementation for other Report properties such as series and
charts.

The patch doesn't break backwards compatibility of the get_baseline method.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>