recipes: add nperf_max_runs alias instead of improper retrieval from confidence parameter
This patch fixes the improper retrieval of maximal number of
iterations when running netperf with confidence level and
interval set. The second number of this parameter specifies
the confidence interval in percents rather than runs.
If a user wants to specify maximal number of iterations he
can do so by specifying nperf_max_runs alias.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Acked-by: Ondrej Lichtner <olichtne@redhat.com>