PerfRepo: add parameter log to REST API methods
This is usefull when we call these methods to just check the result and
don't want to add mess to the logs. For example if a user wants to get a
test specified by ID or UID he first calls the get_by_id method that
fails, but then the get_by_uid succeeds he probably doesn't want the
first attempt to clutter up the logs with errors.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>