add PerfRepoMapping module
This creates a new module that defines a PerfRepoMapping class that
handles PerfRepo mapping files. It uses the Path module to allow for
both locally stored files and http accessible files. Parsing is based on
the utility functions that we used until now, but improved to
(explicitly) allow comments starting with the '#' character, until end
of line.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>