PerfRepoUtils: add template functions for netperf
This adds two new functions that work as templates when working with
Netperf and PerfRepo.

netperf_baseline_template takes a TestModule (Netperf) and a
PerfRepoBaseline object to set the threshold options of the Netperf
module based on data available from the Baseline.

netperf_result_template takes a PerfRepoResult object and result data
returned by a netperf client to add values to the PerfRepoResult object.

These templates are based on how we work with PerfRepo and Netperf so
they might not be useful for everyone.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>