Task: refactor PerfRepo support for Python Tasks
This patch brings brings a complete rehaul of the LNST-PerfRepo
integration as seen and used from Python Tasks. Because of that it
breaks backwards compatibility for recipes that used this functionality.

Most of all the changes use the newly added functionality of
PerfRepoMapping class and the Noop class to remove PerfRepo logic from
recipe tasks.

Other improvements include better logs of what objects are created or
used as baselines and also logs with perfrepo-cli commands to help with
creating reports and setting baselines.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>