PacketCapture: fix capturing into files
This patch fixes the problem with capturing data on configured
interfaces. The problem is that we start tcpdump inside shell. When we
stop the capture we call Popen.terminate() but this terminates the
shell process, not the tcpdump process. Until it's terminated the data are
not flushed to capture file and we transfer empty file to controller.
The fix for this problem is running without shell.

Additionally I'm removing the quotation marks around pcap filter as they
are not required and cause unnecessary entry in log when it is empty.

The last change in the patch is moving redirection of stderr to the code
where Popen is called. This is a must because of using shell=False.

Resolves #150

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Acked-by: Ondrej Lichtner <olichtne@redhat.com>