regression-tests: test only some specific offload combinations
This patch changes which offload combinations we test. Previously we
would test gso, gro and tso in both on and off states - 6 combinations,
and since all the offloads are on by default this means that 3 of those
combinations were identical (gro=on; gso=on; tso=on) resulting in 2
useless tests.

From now on we will only be testing some specific combinations, for now
we have:
* gro=on  + gso=on  + tso=on  + tx=on  + rx=on
* gro=off + gso=on  + tso=on  + tx=on  + rx=on
* gro=on  + gso=off + tso=off + tx=on  + rx=on
* gro=on  + gso=on  + tso=off + tx=off + rx=on
* gro=on  + gso=on  + tso=on  + tx=on  + rx=off

This should halve the test time compared to before. Unfortunately this
patch will change all the hashes generated for PerfRepo results.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>