PerfRepo: remove object initialization from Element objects
The PerfRepo objects have the ability to be initialized from an XML
Element tree, corresponding to an XML response from the REST API.
This was implemented by checking the type of input using the isinstance
method. However on Python 2.6 the API for ElementTree is different and
Element is just a function instead of a class, so this method cannot be
used.

Instead this patch replaces the isinstance call with the iselement()
call that is provided by the ElementTree API in both Python versions.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>