regression-tests: completely remove rx offload
The previous attempt to remove problems with rx offload on team and bond
devices turned out to be incorrect since i just removed the setting
where we tried running rx=off when in fact the problem is that we're
trying to run rx=on in every other setting. This patch removes the rx
offload from all the settings completely to fix this issue.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>