Netperf: add support for multiple threads
The Netperf module now takes an additional option "num_parallel" that
indicates, how many netperf clients should be run in parallel. When the
commands finish, their results are summed into a single result that is
reported back. Using this feature will automatically disable the
confidence feature because this combination doesn't make sense. It is
recommended to use this feature in conjuncture with the parameter "runs"
that will run this setup several times and compute the standard
deviation.

This patch is inspired by super_netperf:
https://github.com/borkmann/stuff/blob/master/super_netperf

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>
Acked-By: Jiri Prochazka <jprochaz@redhat.com>