Netperf: add std_deviation to res_data in case of multiple runs
In case the Netperf client is run multiple times we reported the
standard deviation as just rate deviation and used it to create a
confidence interval around the mean value. Instead we should report the
standard deviation separately and compute the interval as a multiple of
the standard deviation, for now I set it to 2 times the standard
deviation, based on the 68-95-99.7 rule, because it seems comparable to
the intervals reported when -I99 is set for the Netperf client. But we
should investigate this more...

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>
Acked-By: Jiri Prochazka <jprochaz@redhat.com>