Netperf: use "runs" when "confidence" is specified
The "runs" module option was used to specify how many times should the
test module run the netperf client, however it was ignored when
"confidence" was specified since the user was expected to specify the
"-i" option in the task.

After this patch the "runs" parameter will be used even when
"confidence" is specified and will be used to add the "-i" parameter to
the client command.

This patch also updates all the recipes tracked in the LNST recipe that
specified the -i parameter, to instead use the "runs" option.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>