Wizard: fix query of ip address by guest name
Wizard in virtual mode uses libvirt's DHCPLeases() method to check if the
lease table contains guest libvirt domain name. Unfortunately the hostname
key is not the libvirt guest's name but really a hostname. So unless a user
has a mapping in /etc/hosts the current approach won't work and the hostname
contains 'None'. The solution is to get guest's XML and based on the
interface data there search the entry in the lease table provided by
DHCPLeases().

Fixes issue #156

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Acked-by: Jiri Prochazka <jprochaz@redhat.com>
Acked-by: Ondrej Lichtner <olichtne@redhat.com>