Config: pools have their own section
This commit removes the machine_pool_dirs option for the LNST Controller,
and replaces it with a new section - [pools]. In this section you can
specify your machine pools as options like this:
my_pool = ./path/to/my_pool/directory/

This change required some adjustments to the SlavePool module which are
contained in this commit. With that I did some variable renaming so that
it matches the current way of storing pools and I also made small
changes to the logs reported from the SlavePool module so that they make
more sense.

Related to issue #157.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>