lnst-ctl: accept directoris with the --pools option
The --pools option now also accepts a directory path instead of just
pool names. Pool names still have the priority now so if you want to
specify a directory the path you specify can't be the same as the name
of some pool. An easy way to do this is to just include the final '/'
character.

Expands on issue #157

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>