Task: add interface to work with bridge build-in vlans
Kernel exposes an interface primary used by bridge to work with internal
VLAN implementation. However, other devices (for example SR-IOV drivers)
implement this as well making this not bridge-specific.

Allow task to work with the bridge VLAN settings. This is implemented
simply by calling "bridge" utility. This may be later on changed to
pyroute2 implementation, alas it does not implement PF_BRIDGE Netlink
at the moment.

Signed-off-by: Jiri Pirko <jiri@mellanox.com>