Task: add interface to work with bridge FDB entries
Kernel exposes an interface primary used by bridge to work with FDB entries.
However, other devices (for example SR-IOV drivers) implement this as well
making this not bridge-specific.

Allow task to work with the bridge FDB entries. This is implemented
simply by calling "bridge" utility. This may be later on changed to
pyroute2 implementation, alas it does not implement PF_BRIDGE Netlink
at the moment.

Signed-off-by: Jiri Pirko <jiri@mellanox.com>