InterfaceManager: Add speed setting methods
In order to test speed negotiation it is useful to have methods that
can change the speed of an interface.

The first method, 'set_speed', allows one to specify a speed (in Mb/s,
as in ethtool) to be set on an interface. At its lowest level, it's
implemented by calling 'ethtool' with desired 'speed' parameter and
'autoneg' set to 'off'. This is needed in order to prevent the driver
from advertising any other speeds.

On the other end of the link, the 'set_autoneg' method should be used
(again, implemented via 'ethtool'). When set it basically tells the
driver to advertise all the supported speeds.

To make sure no interface is left with auto negotiation off, set auto
negotation on when the Ethernet device is deconfigured.

Note that it's possible to tell the driver which speed and duplex to
advertise in autonegotiation, but for now we omit that and simply
advertise all of them. If needed, this can be added in the future.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Reviewed-by: Jiri Pirko <jiri@mellanox.com>
Signed-off-by: Jiri Pirko <jiri@mellanox.com>