test_modules: Add support for speed test in LinkNeg
It it useful for driver authors to be able to test the speed negotiation
capabilities of their hardware, so add support for it in the LinkNeg
module.

Do that by adding another optional parameter, which represents the
desired speed ('admin_speed'). Test for the desired speed only if the
administrative and operational states match (both should be 'UP') and
report the result accordingly.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Reviewed-by: Jiri Pirko <jiri@mellanox.com>
Signed-off-by: Jiri Pirko <jiri@mellanox.com>