test_modules: introduce "Custom" test module
This is very dump test module. All it knows to do is to get option named
"fail" and according to the value, it either fails or passes. Also, it
takes all options user passed to this test module and pushes them into
result data.

Purpose of this test module is to be used from python tasks, providing
possibility to do "embedded" tests directly inside python tasks.

Signed-off-by: Jiri Pirko <jiri@mellanox.com>
Acked-by: Jan Tluka <jtluka@redhat.com>